Cleanup compile warnings.

+++ jpeginfo.c
@@ -550,7 +550,7 @@ void print_jpeg_info(struct jpeg_info *info)
 	if (csv_mode) {
 		printf("\"%s\",%lu,\"%s\",%d,%d,\"%dbit\",\"%s\",\"%c\",\"%s\",\"%s\",\"%s\",\"%s\"\n",
 			info->filename,
-			info->size,
+			(long unsigned int)info->size,
 			digest,
 			info->width,
 			info->height,
@@ -570,7 +570,7 @@ void print_jpeg_info(struct jpeg_info *info)
 			" \"color_depth\":\"%dbit\", \"type\":\"%s\", \"mode\":\"%s\", \"info\":\"%s\","
 			" \"comments\":\"%s\", \"status\":\"%s\", \"status_detail\":\"%s\" }",
 			info->filename,
-			info->size,
+			(long unsigned int)info->size,
 			digest,
 			info->width,
 			info->height,
@@ -593,7 +593,7 @@ void print_jpeg_info(struct jpeg_info *info)
 		if (longinfo_mode)
 			printf("%-20s ", einfo);
 		printf("%7lu ",
-			info->size);
+			(long unsigned int)info->size);
 		if (info->digest)
 			printf("%s ", digest);
 		if (com_mode)
@@ -616,7 +616,7 @@ void print_jpeg_info(struct jpeg_info *info)
 		if (longinfo_mode)
 			printf("%-20s ", einfo);
 		printf("%7lu ",
-			info->size);
+			(long unsigned int)info->size);
 		if (info->digest)
 			printf("%s ", digest);
 		if (com_mode)
